requirements.txt
alembic==1.0.11
blinker==1.4
boto3==1.9.204
botocore==1.12.204
Click==7.0
docutils==0.14
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Mako==1.1.0
MarkupSafe==1.1.1
Pillow==6.1.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.6
urllib3==1.25.3
Werkzeug==0.15.5
WTForms==2.2.1
